FIXUP for ASoC: Add PCM3060 codec. (incorrect sample format)
pcm3060: Fix incorrect sample format in PCM3060 codec

S24_LE represents samples which consist of 24 data bits and 8 padding bits,
where the padding bits are the *most* significant ones. However, this is
not what the PCM3060 expects; instead, it expects the padding bits to be
the *least* significant ones.

From ALSA's point of view (as well as for most sound chips), there is no
practical difference between true 32 bit samples and
24-bit-with-8-LSB-padding ones; the only difference is that in the latter
case, the 8 LSB are ignored. So, using S32_LE fixes playback.

Signed-off-by: Carlos Rafael Giani <carlos.giani@streamunlimited.com>
Signed-off-by: Radek Dostal <radek.dostal@streamunlimited.com>