watchdog: omap_wdt: Add 'early_disable' module parameter
This parameter makes it possible to control if the watchdog
is being disabled during initialization or will stay enabled
in case it was previously initialized in the bootloader.

To maintain the existing behavior, the default value is 'true',
thus the watchdog is disabled during initialization.

This new feature is highly inspired by the w83627hf_wdt.c

Signed-off-by: Rostislav Lisovy <lisovy@merica.cz>
[set early_disable = true when stopping watchdog, otherwise watchdog
 will not start next time]
Signed-off-by: Radek Dostal <radek.dostal@streamunlimited.com>