ASoC: codecs: add pcm dummy driver.
This driver accepts the specified formats and rates without any hardware
configuration over I2C. It is assumed the DAC is in bootstrap mode or can
auto detect the sample rates and formats received.

The reset gpio defined in the device tree can be used as a hardware reset
signal.

Signed-off-by: Lee Page <lee.page@streamunlimited.com>